requirements.txt
Flask>=2.0.0
python-dotenv==0.19.2
azure-mgmt-storage==19.0.0
azure-identity==1.7.1
seaborn>=0.11.2
